+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".activities.MovieActivity"
+            android:name=".view.activity.MovieActivity"
             android:label="@string/app_name">
             <!--<intent-filter>-->
             <!--<action android:name="android.intent.action.MAIN" />-->
@@ -20,11 +20,11 @@
             <!--</intent-filter>-->
         </activity>
         <activity
-            android:name=".activities.MovieDetailActivity"
+            android:name=".view.activity.MovieDetailActivity"
             android:label="@string/title_activity_movie_detail"
             android:launchMode="singleInstance" />
         <activity
-            android:name=".activities.SettingsActivity"
+            android:name=".view.activity.SettingsActivity"
             android:label="@string/settings_title" />
 
         <provider
@@ -33,15 +33,15 @@
             android:exported="true"
             android:syncable="true" />
 
-        <activity android:name=".view.LoginActivity">
+        <activity android:name=".view.activity.LoginActivity">
             <!--<intent-filter>-->
             <!--<action android:name="android.intent.action.MAIN" />-->
 
             <!--<category android:name="android.intent.category.LAUNCHER" />-->
             <!--</intent-filter>-->
         </activity>
-        <activity android:name=".view.DashBoardActivity" />
-        <activity android:name=".activities.TestActivity">
+        <activity android:name=".view.activity.DashBoardActivity" />
+        <activity android:name=".view.activity.TestActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
