+++ app/src/main/java/com/example/popularmovies/base/data/BaseDatabaseManager.java
@@ -1,4 +1,4 @@
-package com.example.popularmovies.realm;
+package com.example.popularmovies.base.data;
 
 
 public abstract class BaseDatabaseManager<T> implements IDatabaseManager<T> {
