+++ app/src/main/java/com/example/popularmovies/view/activity/DashBoardActivity.java
@@ -1,4 +1,4 @@
-package com.example.popularmovies.view;
+package com.example.popularmovies.view.activity;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
