+++ app/src/main/res/layout/activity_dash_board.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.example.popularmovies.view.DashBoardActivity">
+    tools:context="com.example.popularmovies.view.activity.DashBoardActivity">
 
     <TextView
         android:id="@+id/dashboard_tv_display"
