+++ app/src/main/res/layout/activity_test.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.popularmovies.activities.TestActivity">
+    tools:context="com.example.popularmovies.view.activity.TestActivity">
 
 </android.support.constraint.ConstraintLayout>
